+++ changelog.md
@@ -19,6 +19,13 @@ New
 * User_Agent: detect_operating_system method added
 * User_Agent: list_operating_systems method added
 
+Modified
+----------------
+
+* User_Agent: list_operating_systems method sort names alphabetically
+* User_Agent: list_operating_systems method duplicate operating systems removed
+* User_Agent: list_operating_systems method option to group operating systems
+
 Since php-library v1.5.0 (6-Apr-2018)
 =======